requirements.txt
redis==5.0.1
python-dotenv==1.0.0
openai==1.55.3
httpx==0.27.2